+++ optimize/tests/data/test_struct_chunks.json
@@ -1,14 +1,44 @@
 [
     {
         "text": "lorem ipsum dolor sit amet",
-        "item_id": "abc:1"
+        "item_id": "latin:1",
+        "query_metadata": {
+            "language": "latin"
+        }
     },
     {
         "text": "lorem ipsum relevant chunk super good response",
-        "item_id": "abc:2"
+        "item_id": "latin:2",
+        "query_metadata": {
+            "language": "latin"
+        }
     },
     {
         "text": "lorem ipsum not relevant chunk",
-        "item_id": "abc:3"
+        "item_id": "latin:3",
+        "query_metadata": {
+            "language": "latin"
+        }
+    },
+    {
+        "text": "Mariposa luna callejones sombra azul reloj trigo viento.",
+        "item_id": "spanish:1",
+        "query_metadata": {
+            "language": "spanish"
+        }
+    },
+    {
+        "text": "Sueño estrella bosque río cantar nube fuego piedra hoja.",
+        "item_id": "spanish:2",
+        "query_metadata": {
+            "language": "spanish"
+        }
+    },
+    {
+        "text": "Brisa montaña camino luz copa raíz música silencio alas.",
+        "item_id": "spanish:3",
+        "query_metadata": {
+            "language": "spanish"
+        }
     }
 ]
