+++ optimize/utilities.py
@@ -13,6 +13,36 @@ from optimize.models import EmbeddingSettings, LabeledItem, Settings
 cache_folder = os.getenv("MODEL_CACHE", "models")
 
 
+def schema_from_settings(settings: Settings, additional_schema_fields=None):
+    schema = {
+        "index": {"name": f"{settings.test_id}", "prefix": "rvl"},
+        "fields": [
+            {"name": "text", "type": "text"},
+            {"name": "id", "type": "tag"},
+            {"name": "file_name", "type": "tag"},
+            {"name": "item_id", "type": "tag"},
+            {
+                "name": "vector",
+                "type": "vector",
+                "attrs": {
+                    "dims": settings.embedding.dim,
+                    "distance_metric": settings.index.distance_metric,
+                    "algorithm": settings.index.algorithm,
+                    "datatype": settings.index.vector_data_type,
+                    "ef_construction": settings.index.ef_construction,
+                    "ef_runtime": settings.index.ef_runtime,
+                    "m": settings.index.m,
+                },
+            },
+        ],
+    }
+
+    if additional_schema_fields:
+        for field in additional_schema_fields:
+            schema["fields"].append(field)
+    return schema
+
+
 def load_labeled_items(settings: Settings):
     with open(settings.data.labeled_data_path, "r") as f:
         labeled_items = json.load(f)
